# Improve performance of `Enumerable.ToDictionary()`