# Summary

In this Pull-Request, I've improved the performance of the `Enumerable.ToDictionary()` method in the following two points.

1. Set the initial capacity when the source implements `ICollection` or `IIListProvider<T>`
    - The implementation patterns where the initial capacity is set are increasing
2. Combine the specializations for `T[]` and `List<T>` into a specialization for `ReadOnlySpan<T>`
    - Reduction of code volume
    - Particularly, speed up of enumeration for `List<T>`


# Benchmark

```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22621.2861/22H2/2022Update/SunValley2)
13th Gen Intel Core i7-1360P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.100
  [Host]   : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2
  ShortRun : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1
WarmupCount=3
```

| Method                    | Mean     | Error     | StdDev   | Ratio | RatioSD | Gen0     | Gen1     | Gen2     | Allocated | Alloc Ratio |
|-------------------------- |---------:|----------:|---------:|------:|--------:|---------:|---------:|---------:|----------:|------------:|
| EnumerableToDictionary    | 236.9 us |  29.52 us |  1.62 us |  1.00 |    0.00 | 153.8086 | 153.8086 | 153.8086 | 788.75 KB |        1.00 |
| ArrayToDictionary         | 110.9 us |  16.13 us |  0.88 us |  0.47 |    0.01 |  62.3779 |  62.3779 |  62.3779 | 236.94 KB |        0.30 |
| ListToDictionary          | 136.4 us | 660.76 us | 36.22 us |  0.58 |    0.15 |  62.3779 |  62.3779 |  62.3779 | 236.94 KB |        0.30 |
| PR_EnumerableToDictionary | 170.8 us |  19.04 us |  1.04 us |  0.72 |    0.01 |  62.2559 |  62.2559 |  62.2559 | 236.99 KB |        0.30 |
| PR_ArrayToDictionary      | 114.8 us |  20.21 us |  1.11 us |  0.48 |    0.00 |  62.3779 |  62.3779 |  62.3779 | 236.94 KB |        0.30 |
| PR_ListToDictionary       | 115.1 us |  20.52 us |  1.12 us |  0.49 |    0.01 |  62.3779 |  62.3779 |  62.3779 | 236.94 KB |        0.30 |


```cs
public class ToDictionaryBenchmark
{
    private const int Length = 10000;

    private static readonly IEnumerable<Kvs> s_enumerable = Enumerable.Range(0, Length).Select(x => new Kvs(x, x));
    private static readonly IEnumerable<Kvs> s_array = s_enumerable.ToArray();
    private static readonly IEnumerable<Kvs> s_list = s_enumerable.ToList();

    [Benchmark(Baseline = true)]
    public Dictionary<int, Kvs> EnumerableToDictionary()
        => s_enumerable.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> ArrayToDictionary()
        => s_array.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> ListToDictionary()
        => s_list.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_EnumerableToDictionary()
        => s_enumerable.ToDictionary_PR(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_ArrayToDictionary()
        => s_array.ToDictionary_PR(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_ListToDictionary()
        => s_list.ToDictionary_PR(static x => x.Key);
}

public record struct Kvs(int Key, int Value);
```