@stephentoub 2024-01-06
Thanks. What did the comparison look like, in particular for the enumerable one, with a small length, like 3?