@xin9le 2024-01-08
@stephentoub 
Thank you for your approval. As you already understand, I believe there may be slight overhead in the following cases:

- When size information cannot be obtained
- When the internal buffer of the collection type is not expanded

※ This is due to the increased type determination in `.TryGetNonEnumeratedCount()`

Whether or not this overhead is considered negligible compared to the benefits gained from setting the initial capacity, I believe, is the determining factor in whether or not to incorporate this Pull-Request. Although you have already approved, I will document the measurement results just in case.

# Benchmark

```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22621.2861/22H2/2022Update/SunValley2)
13th Gen Intel Core i7-1360P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.100
  [Host]     : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

| Method                    | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| EnumerableToDictionary    | 39.45 ns | 0.817 ns | 1.431 ns |  1.00 |    0.00 | 0.0263 |     248 B |        1.00 |
| ArrayToDictionary         | 34.26 ns | 0.698 ns | 0.776 ns |  0.87 |    0.05 | 0.0221 |     208 B |        0.84 |
| ListToDictionary          | 34.20 ns | 0.570 ns | 0.701 ns |  0.87 |    0.04 | 0.0221 |     208 B |        0.84 |
| PR_EnumerableToDictionary | 41.02 ns | 0.814 ns | 0.762 ns |  1.05 |    0.07 | 0.0263 |     248 B |        1.00 |
| PR_ArrayToDictionary      | 34.18 ns | 0.695 ns | 0.543 ns |  0.88 |    0.05 | 0.0221 |     208 B |        0.84 |
| PR_ListToDictionary       | 32.78 ns | 0.683 ns | 0.670 ns |  0.84 |    0.05 | 0.0221 |     208 B |        0.84 |
```

```cs
public class ToDictionaryBenchmark
{
    private static readonly IEnumerable<Kvs> s_enumerable = EnumerateSample();
    private static readonly IEnumerable<Kvs> s_array = s_enumerable.ToArray();
    private static readonly IEnumerable<Kvs> s_list = s_enumerable.ToList();

    private static IEnumerable<Kvs> EnumerateSample()
    {
        for (var i = 0; i < 3; i++)
            yield return new Kvs(i, i);
    }

    [Benchmark(Baseline = true)]
    public Dictionary<int, Kvs> EnumerableToDictionary()
        => s_enumerable.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> ArrayToDictionary()
        => s_array.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> ListToDictionary()
        => s_list.ToDictionary(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_EnumerableToDictionary()
        => s_enumerable.ToDictionary2(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_ArrayToDictionary()
        => s_array.ToDictionary2(static x => x.Key);

    [Benchmark]
    public Dictionary<int, Kvs> PR_ListToDictionary()
        => s_list.ToDictionary2(static x => x.Key);
}

public record struct Kvs(int Key, int Value);
```